requirements.txt
PySide6>=6.2
opencv-python>=4.8.0
pytesseract
Pillow